pyproject.toml
[project]
name = "geo-guard-kenya"
version = "0.1.0"
description = "Real-time disaster early warning system for Floods, Droughts, and Landslides."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi>=0.128.0",
    "folium>=0.20.0",
    "geoalchemy2>=0.18.1",
    "geopandas>=1.1.2",
    "google-genai>=1.62.0",
    "joblib>=1.5.3",
    "numpy>=2.4.1",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "plotly>=6.5.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "shapely>=2.1.2",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.53.0",
    "streamlit-folium>=0.26.1",
    "twilio>=9.10.1",
    "uvicorn>=0.40.0",
]
